requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
keras
opencv-python
Pillow